+++ src/IpTraceable/Mapping/Driver/Xml.php
@@ -63,7 +63,10 @@ class Xml extends BaseXml
                     if (!$this->_isAttributeSet($data, 'on') || !in_array($this->_getAttribute($data, 'on'), ['update', 'create', 'change'], true)) {
                         throw new InvalidMappingException("Field - [{$field}] trigger 'on' is not one of [update, create, change] in class - {$meta->getName()}");
                     }
-
+                    if ($this->_isAttributeSet($data, 'setterMethod')) {
+                        $setterMethod = $this->_getAttribute($data, 'setterMethod');
+                        $this->setSetterMethod($field, $setterMethod, $config);
+                    }
                     if ('change' === $this->_getAttribute($data, 'on')) {
                         if (!$this->_isAttributeSet($data, 'field')) {
                             throw new InvalidMappingException("Missing parameters on property - {$field}, field must be set on [change] trigger in class - {$meta->getName()}");