+++ CHANGELOG.md
@@ -27,6 +27,7 @@ a release.
 - Sluggable: Allow ascii_string to validTypes
 - IpTraceable: Allow ascii_string to validTypes
 - Sluggable: Use `TranslationWalker` hint when looking for similar slugs (`getSimilarSlugs` method) for entities which implement `Translatable` interface and have `uniqueOverTranslations: true` Slug option (#100, #2530)
+- Blameable/IpTraceable/SoftDeletable/Timestampable: Added functionality to use setter method instead of setting property values directly (#2644)
 
 ## [3.15.0]
 ### Added